refactor(all): move `deepMerge`  from `@agent-tars/core` to `@multimodal/shared-utils`
