+++ multimodal/agent-tars/package.json
@@ -22,6 +22,7 @@
     "benchmark:crawl": "cd benchmark/crawl && pnpm start"
   },
   "dependencies": {
+    "@multimodal/shared-utils": "workspace:*",
     "@multimodal/shared-media-utils": "workspace:*"
   },
   "devDependencies": {
